+++ requirements.txt
@@ -1,0 +1,8 @@
+# This file may be used to create an environment using:
+# $ conda create --name <env> --file <this file>
+# platform: linux-64
+numpy=1.25.2=py39h6183b62_0
+python=3.9.18=h0755675_0_cpython
+rdkit=2023.9.5=pypi_0
+rdkit-pypi=2022.9.5=pypi_0
+scipy=1.11.2=py39h474f0d3_1
